requirements.txt
hydra-core==1.3.2
numpy==1.25.2
omegaconf==2.3.0
pandas==2.1.2
scikit_learn==1.2.2
tqdm==4.65.0
xgboost==1.7.3
optuna==3.4.0
scipy==1.11.3
matplotlib==3.8.0
fiona==1.9.1
seaborn==0.12.2
xgboost==1.7.3
shap==0.42.1
tqdm==4.65.0
gdown==4.7.1
flake==6.1.0
omegaconf==2.3.0
dask==2023.6.0
openpyxl==3.0.10
pyarrow==11.0.0
catboost==1.2.2
black
nbstripout